CASH-402: Refactor BorrowerInfo to increase shared code
